feat(clickhousech@4135480): set plan semgent pipeline threads independent of exchange_parallel_size [skipped]

See merge request dp/ClickHouse!5554

+++ src/Processors/Exchange/ExchangeUtils.h
@@ -92,7 +92,7 @@ public:
             total_memory_tracker.alloc(bytes);
     }
 
-    static inline void transferGlobalMemoryToThread(size_t bytes)
+    static inline void transferGlobalMemoryToThread(Int64 bytes)
     {
         if (DB::MainThreadStatus::get())
             total_memory_tracker.free(bytes);
